project/requirements.txt
urllib3
certifi
pyjwt
pyjwt[crypto]
asyncio
SQLAlchemy